Util関数をrenderer/util配下に移動、class化、class名もより汎用的な名称に変更
